+++ packages/xc-cfg/src/configs/solana/solana.ts
@@ -1,6 +1,6 @@
 import { AssetRoute, ChainRoutes } from '@galacticcouncil/xc-core';
 
-import { jito_sol, sol } from '../../assets';
+import { jito_sol, prime, sol } from '../../assets';
 import { solana, hydration, moonbeam } from '../../chains';
 import { BalanceBuilder, ProgramBuilder } from '../../builders';
 import { Tag } from '../../tags';
@@ -51,6 +51,36 @@ const toHydrationViaWormhole: AssetRoute[] = [
       fee: {
         amount: 0,
         asset: jito_sol,
+      },
+    },
+    program: ProgramBuilder()
+      .Wormhole()
+      .TokenBridge()
+      .transferTokenWithPayload()
+      .viaMrl({
+        moonchain: moonbeam,
+      }),
+    tags: [Tag.Mrl, Tag.Wormhole],
+  }),
+  new AssetRoute({
+    source: {
+      asset: prime,
+      balance: BalanceBuilder().solana().token(),
+      fee: {
+        asset: sol,
+        balance: BalanceBuilder().solana().native(),
+      },
+      destinationFee: {
+        asset: prime,
+        balance: BalanceBuilder().solana().token(),
+      },
+    },
+    destination: {
+      chain: hydration,
+      asset: prime,
+      fee: {
+        amount: 0,
+        asset: prime,
       },
     },
     program: ProgramBuilder()