+++ packages/xc-sdk/src/platforms/substrate/SubstrateClaim.ts
@@ -10,6 +10,8 @@ import { encoding } from '@wormhole-foundation/sdk-base';
 
 import { encodeFunctionData } from 'viem';
 
+import { Binary } from 'polkadot-api';
+
 import { SubstrateCall } from './types';
 import { EvmCall } from '../evm';
 
@@ -47,18 +49,28 @@ export class SubstrateClaim {
     const api = client.getUnsafeApi();
     const claim = this.redeemMrl(from, vaaBytes);
     const tx = (api.tx as any).EthereumXcm.transact({
-      V2: {
-        gas_limit: 5_000_000n,
-        action: {
-          Call: claim.to,
+      xcm_transaction: {
+        type: 'V2',
+        value: {
+          gas_limit: [5_000_000n, 0n, 0n, 0n],
+          fee_payment: {
+            type: 'Auto',
+            value: undefined,
+          },
+          action: {
+            type: 'Call',
+            value: Binary.fromHex(claim.to),
+          },
+          value: [0n, 0n, 0n, 0n],
+          input: Binary.fromHex(claim.data),
+          access_list: undefined,
         },
-        value: 0n,
-        input: claim.data,
       },
     });
 
+    const encoded = await tx.getEncodedData();
     return {
-      data: tx.decodedCall,
+      data: encoded.asHex(),
       from: from,
       type: CallType.Substrate,
       dryRun: async () => undefined,